refactor: split Hina.Host into testable units, add Hina.Host.Tests
Program.cs (535 lines, all classes inline, untested) is now wiring only;
HostOptions, Routing, AccessStats and SetupWizard moved to their own
files under namespace Hina.Host. The wizard's config-file check is
separated from the Console.IsInputRedirected gate so it is unit-testable.

New Hina.Host.Tests (25 tests): HostOptions load/override matrix,
Routing, AccessStats counters + abuse throttling, wizard config probe,
and in-process endpoint tests via WebApplicationFactory.

Bug found by the endpoint tests: static files never served *.chunk.br —
the default content-type provider has no .br mapping, so every chunk
request returned 404. Fixed with ServeUnknownFileTypes +
application/octet-stream default.
